+++ libsolidity/ast/ASTJsonImporter.cpp
@@ -1004,10 +1004,6 @@ Literal::SubDenomination ASTJsonImporter::subdenomination(Json::Value const& _no
 		return Literal::SubDenomination::Wei;
 	else if (subDenStr == "gwei")
 		return Literal::SubDenomination::Gwei;
-	else if (subDenStr == "szabo")
-		return Literal::SubDenomination::Szabo;
-	else if (subDenStr == "finney")
-		return Literal::SubDenomination::Finney;
 	else if (subDenStr == "ether")
 		return Literal::SubDenomination::Ether;
 	else if (subDenStr == "seconds")